Move altitude stuff to its own file
